backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
alembic
python-dotenv
psycopg2-binary
python-pptx
